requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
anaconda-anon-usage 
annotated-types==0.7.0
anyio==4.8.0
archspec
asgiref==3.8.1
asttokens==3.0.0
attrs==25.1.0
boltons 
Brotli 
certifi 
cffi 
charset-normalizer 
colorama==0.4.6
comm==0.2.2
conda 
conda-anaconda-telemetry 
conda-content-trust 
conda-libmamba-solver 
conda-package-handling 
conda_package_streaming 
contourpy==1.3.1
crispy-bootstrap4==2023.1
cryptography 
cycler==0.12.1
debugpy==1.8.11
decorator==5.1.1
distro 
Django==5.1.6
django-cors-headers==4.7.0
django-crispy-forms==2.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
executing==2.1.0
filelock==3.17.0
fonttools==4.56.0
frozendict
frozenlist==1.5.0
fsspec==2025.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna 
ipykernel==6.29.5
ipython==8.31.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
jsonpatch 
jsonpointer==2.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
libmambapy 
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
menuinst 
mongoengine==0.29.1
mpmath==1.3.0
multidict==6.1.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.1.1
openai==0.28.0
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
parso==0.8.4
pillow==11.1.0
platformdirs==4.3.6
pluggy 
prompt_toolkit==3.0.48
propcache==0.2.1
psutil==6.1.1
psycopg2==2.9.10
pure_eval==0.2.3
py-cpuinfo==9.0.0
pycosat 
pycparser 
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.18.0
PyJWT==2.9.0
pymongo==3.11.4
pyparsing==3.2.1
PySocks
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.0
requests 
ruamel.yaml
ruamel.yaml.clib 
scipy==1.15.2
seaborn==0.13.2
setuptools==75.1.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
stack-data==0.6.3
sympy==1.13.1
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm 
traitlets==5.14.3
truststore
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.81
ultralytics-thop==2.0.14
urllib3 
wcwidth==0.2.13
wheel==0.44.0
win-inet-pton 
yarl==1.18.3
zstandard 
